Move Text.h over to i18n too
Never use relative paths for #includes. Didn't touch the rest because
out of scope.
